+++ crates/bedit-cloudburst/src/bin/main.rs
@@ -1,5 +1,6 @@
 use bedit_cloudburst::Torrent;
 use std::{
+    error::Error,
     fs::File,
     io::{BufReader, Read},
     path::Path,
@@ -42,24 +43,30 @@ fn check_torrents(path: &Path) -> Result<Vec<Torrent>, serde_bencode::Error> {
 }
 
 fn main() {
+    pretty_env_logger::init();
+
     for arg in std::env::args().skip(1) {
         let path = Path::new(&arg);
         if path.is_file() {
             match torrent_from_file(path) {
                 Ok(torrent) => {
-                    dbg!(torrent);
+                    println!("Deseralized torrent: {}", torrent.name());
                 }
-                Err(e) => eprintln!("Torrent failed to deserialize: {path:?}\n{e}")
+                Err(e) => eprintln!(
+                    "Torrent failed to deserialize: {path:?}\nError: {e}, Error source: {}",
+                    e.source()
+                        .map_or_else(|| "No source".to_string(), |e| e.to_string())
+                ),
             }
-        }
-        else if path.is_dir() {
+        } else if path.is_dir() {
             match check_torrents(path) {
                 Ok(torrents) => {
                     for torrent in torrents {
-                        dbg!(torrent);
+                        println!("Deseralized torrent: {}", torrent.name());
                     }
                 }
-                Err(e) => eprintln!("Failed: {path:?}\n{e}")
+                Err(e) => eprintln!("Failed deserializing directory of torrents: {path:?}\nError: {e}\nError source: {}", e.source().map_or_else(|| "No source".to_string(), |e| e.to_string())),
+
             }
         }
     }