+++ crates/bedit-cloudburst/src/lib.rs
@@ -1,18 +1,17 @@
 //! `bedit-cloudburst` provides strongly typed data structures for serializing and deserializing torrents.
 
-mod crypto;
+pub mod crypto;
 mod fileattributes;
 mod files;
-mod hex;
+pub mod hexadecimal;
 mod info;
 mod pieces;
 mod signature;
 mod torrent;
+mod urlwrapper;
 
-pub use crypto::Sha1Hash;
 pub use fileattributes::{FileAttribute, TorrentFileAttributes};
 pub use files::{FileTree, FileTreeEntry, FileTreeInfo, SharedFiles};
-pub use hex::{HexBytes, Hexadecimal, PackedHex};
 pub use info::{Hybrid, Info, MetaV1, MetaV2};
 pub use pieces::{PieceLength, Pieces};
 pub use signature::{SignInfo, Signature};