+++ crates/bedit-cloudburst/src/pieces.rs
@@ -1,18 +1,18 @@
-use log::error;
+use crate::{crypto::sha1::Sha1Hash, hexadecimal::HexBytes};
+use log::{debug, error};
 use serde::{
     de::{Error as DeError, Unexpected},
     Deserialize, Deserializer, Serialize,
 };
-use serde_bytes::ByteBuf;
 use std::num::NonZeroU64;
 
-use crate::crypto::Sha1Hash;
+const PIECES_DE_TARGET: &str = "bedit_cloudburst::Piece::deserialize";
+const PIECELENGTH_DE_TARGET: &str = "bedit_cloudburst::PieceLength::deserialize";
 
 /// Number of bytes per piece.
 ///
 /// According to the spec, piece length should be greater than 16 KiB and is always a power of two.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Serialize)]
-#[repr(transparent)]
 #[serde(transparent)]
 pub struct PieceLength(NonZeroU64);
 
@@ -21,12 +21,17 @@ impl<'de> Deserialize<'de> for PieceLength {
     where
         D: Deserializer<'de>,
     {
+        debug!(target: PIECELENGTH_DE_TARGET, "Deserializing PieceLength.");
+
         let piece_length = NonZeroU64::deserialize(deserializer)?;
 
         if piece_length.get() >= 16 && piece_length.is_power_of_two() {
             Ok(PieceLength(piece_length))
         } else {
-            error!(target: "bedit_cloudburst::PieceLength::deserialize", "Invalid piece length: {piece_length}.");
+            error!(
+                target: PIECELENGTH_DE_TARGET,
+                "Invalid piece length: {piece_length}."
+            );
             Err(DeError::invalid_value(
                 Unexpected::Unsigned(piece_length.into()),
                 &"piece length should be greater than 16 and a power of two",
@@ -39,24 +44,37 @@ impl<'de> Deserialize<'de> for PieceLength {
 ///
 /// The BitTorrent spec specifies `pieces` as a single byte buffer that is a multiple of 20.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize)]
-#[repr(transparent)]
 #[serde(transparent)]
-pub struct Pieces(ByteBuf);
+pub struct Pieces(HexBytes);
 
 impl<'de> Deserialize<'de> for Pieces {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
     where
         D: Deserializer<'de>,
     {
-        // This is already a byte string so I can't really validate it.
-        let pieces = ByteBuf::deserialize(deserializer)?;
+        debug!(target: PIECES_DE_TARGET, "Deserializing Pieces.");
 
-        // `pieces` must be a multiple of 20.
+        // This is already a byte string so I can't really validate it.
+        let pieces: HexBytes = match serde_bytes::ByteBuf::deserialize(deserializer) {
+            Ok(bytes) => HexBytes::new(bytes),
+            Err(e) => {
+                error!(
+                    target: PIECES_DE_TARGET,
+                    "Error deserializing the byte string for `Pieces`.\nError: {e}"
+                );
+                return Err(e);
+            }
+        };
+
+        // `pieces` must be a multiple of 20 because they're SHA-1 hashes.
         let len = pieces.len();
         if len % 20 == 0 {
             Ok(Pieces(pieces))
         } else {
-            error!(target: "bedit_cloudburst::Piece::deserialize", "Pieces should be a multiple of twenty; got: {len}");
+            error!(
+                target: PIECES_DE_TARGET,
+                "Pieces should be a multiple of twenty; got: {len}"
+            );
             Err(DeError::invalid_length(
                 len,
                 &"length of `pieces` should be a multiple of 20",
@@ -69,7 +87,7 @@ impl Pieces {
     /// Iterator over chunks of 20 bytes.
     #[inline]
     pub fn iter_pieces_bytes(&self) -> impl Iterator<Item = &[u8]> + '_ {
-        self.0.chunks_exact(20)
+        self.0.as_slice().chunks(20)
     }
 
     /// Iterator over bytes wrapped in [Sha1Hash].